flask-ai/requirements.txt
--prefer-binary
chromadb==1.0.13


Flask==3.0.3
Flask-Cors==5.0.0

fpdf==1.7.2

google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-generativeai==0.8.4





llama-cloud==0.1.26
llama-cloud-services==0.6.32
llama-index==0.12.42
llama-index-agent-openai==0.4.11
llama-index-cli==0.4.3
llama-index-core==0.12.42
llama-index-embeddings-huggingface==0.5.4
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.7.7
llama-index-llms-ollama==0.6.2
llama-index-llms-openai==0.4.5
llama-index-multi-modal-llms-openai==0.5.1
llama-index-program-openai==0.3.2
llama-index-question-gen-openai==0.3.1
llama-index-readers-file==0.4.9
llama-index-readers-github==0.6.1
llama-index-readers-llama-parse==0.4.0
llama-index-vector-stores-chroma==0.4.2
llama-parse==0.6.32
lxml==5.3.2
lxml_html_clean==0.4.2
Markdown==3.6
markdown-it-py==3.0.0
ollama==0.5.1



pybase64==1.4.1

pydub==0.25.1
PyGithub==2.6.1

pymongo==4.11.1
PyMuPDF==1.25.5

pyparsing==3.1.2
pypdf==5.6.0
PyPDF2==3.0.1

python-docx==1.1.2
python-dotenv==1.1.0

regex==2024.5.15
requests==2.31.0


sentence-transformers==4.0.2



transformers==4.42.3



